refactor: after mentors review
