+++ src/App.tsx
@@ -2,11 +2,9 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import MainComponent from './components/MainComponent/MainComponent';
 import UncontrolledForm from './components/UncontolledForm/UncontrolledForm';
 import { Provider } from 'react-redux/es/exports';
-import { setupStore } from './stores/store';
+import { store } from './stores/store';
 import ControlledForm from './components/ControlledForm/ControlledForm';
 import './App.css';
-
-const store = setupStore();
 
 const App = () => (
   <Provider store={store}>
@@ -21,7 +19,7 @@ const App = () => (
           element={<UncontrolledForm />}
         />
         <Route
-          path="/react-Hook-Form"
+          path="/react-hook-form"
           element={<ControlledForm />}
         />
       </Routes>
