+++ src/stores/store.ts
@@ -18,6 +18,8 @@ export const setupStore = (preloadedState?: PreloadedState<RootState>) => {
   });
 };
 
+export const store = setupStore();
+
 export type RootState = ReturnType<typeof rootReducer>;
 export type AppStore = ReturnType<typeof setupStore>;
 export type AppDispatch = AppStore['dispatch'];
